Add NOAA Cloud sites (#110)
Add AWS, Azure, and Google Cloud platforms.

They all have an identical filesystem with identical compilers so the only difference between them is the external packages. Updated to the v2 images.

Weird problem with libxml2 that fails to untar on Azure, so I set it to be external
gftl-shared fails to download submodule (perhaps because of ancient 1.8 git), so use .tar.gz releases
v2 image of AWS identifies as centos7 instead of amzn2.