Add jedi-ewok app (jedi-ewok-env with dependency on ecflow, cartopy, ...) (#120)
Add jedi-ewok app and corresponding jedi-ewok-env virtual environment package. This entails

Manually installing qt@5 on several HPC platforms (gaea, discover), because building it in spack is a nightmare (only really builds with gcc, and is also so deeply embedded in the system/X libraries that it comes with a huge number of X packages), and adding it as external package
Installing it via homebrew on macOS and adding as external package
The new ecflow dependency requires building some of the boost libraries
Adding many more modules to the blacklist(s)
Additional changes:

Update top-level README.md with temporary install locations on a number of HPCs for testing - also with information on where the new jedi-ewok app has been tested
The back and forth: remove the jedi-ufs-all app and virtual environment, since we don't want to use it in CI, it increases the changes for duplicates (installing one app first and then adding the dependencies later with --reuse works better), and its name is entirely confusing (because there is already a jedi-ufs app)
Clean up how MPI (openmpi, mpich) is found in .github/actions/setup-spack-stack/action.yaml
Temporarily point to spack branch bugfix/hash_length_container in @climbfuji's fork (contains several updates for building Python modules with gcc on macOS, and a bug fix for defining the hash length for modules needed for future container work)
Remove incorrect dependencies on git-lfs and update how fckit is built (better way to handle C++ libraries suggested by ECMWF developers)
Additional updates on macOS, such as installing basic libraries (curl, wget, ...) with brew
Update of meta_modules/setup_meta_modules.py: For some environments, there are only compiler+mpi-dependent modules, and therefore the compiler itself is not recorded in compiler_dict. This PR fixes it

Co-authored-by: Kyle Gerheiser <3209794+kgerheiser@users.noreply.github.com>