Harden CI (#1143)
Harden CI scripts so that they can be used on a newly provisioned system for the first and all subsequent spack-stack builds (after the prerequisites were installed, per instructions).

Also applies long overdue updates to the macos CI workflow so that it looks closer to the Ubuntu workflows. Note that we can't use the padded_length logic due to errors of invalid mach-o headers when modifying the install path length during the spack relocation.