Update Nautilus site config and common packages/modules to support oneAPI compilers (icx, icpx, ifort) (#1135)
This PR adds a oneapi@2024.1.2 option to the nautilus site config. This option currently uses intel-oneapi-mpi instead of openmpi as MPI provider. If in-depth testing over the next weeks shows that this doesn't perform on Nautilus, we can change it.

The PR further adds the missing append-path logic to the stack-* meta modules, sets the I_MPI_* environment variables in the MPI meta modules (it doesn't hurt if those are always set, other MPI providers ignore them), and updates the documentation for supported compilers.