Support multiple esmf versions in unified environment / updates of neptune-dev, unified-dev, skylab-dev templates (add `%oneapi`) (#1151)
This PR makes it possible to use different ESMF versions for virtual packages within one large environment. NEPTUNE needs esmf@8.7.0b04 (for now), while the other applications need esmf@8.6.1.

There are also cosmetic (indentation) changes for the skylab-dev and unified-dev templates, and an update of the neptune-dev template. I've turned off the Python variant for neptune-env when built as part of the unified environment due to numerous version conflicts and duplicates. We'll want to look at those later, after #1138 was merged and after the py-xnrl package dependencies are updated (ongoing at NRL).