Separate site configs into tier-1 and tier-2 platforms (#1186)
This PR splits up the site configs into subdirectories tier1 and tier2. This is done because the number of sites is getting longer, and it makes it easier to identify which ones are fully supported and which aren't. Further, we can then add more site configs in tier2, for example as needed by spack-stack developers.