Remove unused options from "spack stack" extension and add logic to set compiler for "spack stack create env" (#1192)
This PR removes unused options for the "spack stack" extension in an order to clean up the amount of code and facilitate adding the new capability described below.

The new capability requires the user to define a compiler --compiler when running spack stack create env. It further updates the logic for creating common and site environment spack config files for modules and packages. For both common and site configs (in configs/common/ and configs/sites/tier?/*/), the following is true:

- For packages: If both packages.yaml and packages_${COMPILER}.yaml exist in the same directory, merge the two (packages_${COMPILER}.yaml overrides packages.yaml if necessary) and write to the environment's common or site config as packages.yaml. See for example the updated blackpearl site config in this PR. If only one of them exists, copy that file to the destination.
- For modules: If both modules.yaml and packages_${LUA_OR_LMOD}.yaml exist in the same directory, merge the two (packages_${LUA_OR_LMOD}.yaml overrides modules.yaml if necessary) and write to the environment's common or site config as modules.yaml. This is currently not the case, but maybe we can consolidate some of the contents into one default modules.yaml file?

In the environment spack.yaml, update the compiler definition line for the matrix (if applicable) and set packages:all:require:['%COMPILER']. This requires that packages that must be built with other compilers (e.g. gcc-runtime) have their own require: '%COMPILER' setting to overwrite this. This is already the case for gcc-runtime in configs/common/packages.yaml. The same may be needed for other runtime packages for each of the compilers in the future, and for packages like bison that don't build with oneapi for example. I am basically following what E4S does.