"Revert" Intel deprecation flag changes in spack's lib/spack/env/cc and pull in improved solution from spack mainline (#1238)
This PR resolves #1230. Detailed information on what this PR does and why it is necessary is given in the issue. The improved solution does away with having to specify the diag-disable flags, since these are now set implicitly (and only during spack environment builds, not in downstream applications). Credits to Peter Scheibel from the spack core team for suggesting it and adding the spack unit tests to my implementation.