Add 'py-dask +delayed' to gmao-swell-env and add comments in that package and in py-xnrl about the +delayed variant (#1246)
Add 'py-dask +delayed' to gmao-swell-env (so that it is in the unified environment, as requested by GMAO originally) and add comments in that package and in py-xnrl about the +delayed variant. Because this variant affects which versions of py-dask can be used, we should modify both packages at the same time. At the moment, we still need the +delayed variant in py-xnrl.