Redefine variants in neptune-env virtual package (#1265)
This PR defines three variants in the neptune-env virtual package: espc, python, xnrl. In the unified environment and in the skylab environment, the variants espc and xnrl are off, the python variant is on. This is so that py-fortranformat is available in those environments (required by pyiri-jedi). In the neptune standalone environment, all three variants are on. Note that the xnrl variant requires the python variant.