Enable modules for packages not compiled with preferred compiler (update spack-stack setup-meta-modules); remove need for external `ecflow` (#1257)
This PR makes the necessary changes to the site configs and the setup-meta-modules extension to build certain packages with other compilers than the preferred compiler and still be able to load the module for that package. We can only do this now that we have the concept of a preferred compiler in our environments. One assumption made here is that the MPI provider, if any, is compiled with the preferred compiler (I think this is a reasonable assumption to make).

Our use cases are:

bison needs to be compiled with gcc when the preferred compiler is oneapi. Strictly speaking, we don't need the bison module in this case, but I tested it on my laptop and it works.
ecflow and boost must be compiled with gcc when the preferred compiler is intel. This allows us to move away from external ecflow packages that don't work with the proposed update of Python to 3.11.7 (because the external ecflow was compiled with an old Python 3.9). In this case, we need the ecflow modulefile. I tested this on the Ubuntu CI runner and on Narwhal.
Caveat: I have not tested if this new capability works with packages that depend on MPI (which is compiled with the preferred compiler) but that get compiled with a different compiler (e.g. something like intel-oneapi-mpi/2021.12.0/gcc/11.2.0 when the packages using the preferred compiler would have intel-oneapi-mpi/2021.12.0/intel/2021.12.0).

Still todo:
- Configure gcc compiler used as backend for Intel for Atlantis, Gaea C5, Gaea C6 (deferred to follow-up PRs)
- Remove external ecflow from packages.yaml and make sure every site has an external qt@5 in its packages.yaml