Pin matplotlib and numpy to 3.7.4 and 1.23.5 (#1283)
During testing of an early 1.8.0 candidate, it was found that compiling numpy v1.25.x and matplotlib v3.8.x together with Intel caused matplotlib to fail to generate basic plots that included tick marks. Reverting the versions to 1.23.5 and 3.7.4 resolved the issue.

This version does not work on Acorn with Intel 19, but other than for newer versions of Intel, numpy-1.24.4 works there. Therefore, there will be a follow-up PR that changes the default py-numpy version on Acorn for Intel 19 only to 1.24.4.