release/1.8.0: bug fix for compiling g2@3.5.1 with gfortran < 10 in spack package, remove from container configs (#1287)
For release/1.8.0: Update submodule pointer for spack for the changes in JCSDA/spack#477. Remove g2 compiler flags in container configs, since these are now in the g2 package itself.

Also include a bug fix to adjust the py-numpy version to match what is in configs/common/packages.yaml since yesterday (1.23.5).