Allow specifying compiler version for 'spack stack create env' (#1294)
This PR borrows the logic from #1213 that allows a compiler version to be specified in spack stack create env. It's a couple of tweaks to stack_env.py. This isn't strictly urgent, but I'm hoping to put this in the 1.8.0 release in anticipation of supporting installations by NCO (plus maybe others will find it useful; I know it will be useful for Acorn, and maybe also for NRL?).